requirements.txt
ollama
pywebview[gtk]
chromadb
ddgs
